Reorganize landing page component structure
