+++ web-app/src/app/(landing)/how-it-works.tsx
